upd: refactor tomb -> errgroup and context
